requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
beautifulsoup4>=4.12.0
groq>=0.4.0
pandas>=2.0.0
matplotlib>=3.7.0
rich>=13.0.0
python-dotenv>=1.0.0
